+++ apps/ced-portal-be/CHANGELOG.md
@@ -1,5 +1,27 @@
 # ced-portal-be
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [89fa17c]
+  - @pagopa/io-core-adapter-tracing@0.0.6
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [13a5a52]
+  - @pagopa/io-core-adapter-tracing@0.0.5
+
+## 0.1.1
+
+### Patch Changes
+
+- 1e199aa: manage non json replies and add admin auth on department api
+- Updated dependencies [1e199aa]
+  - @pagopa/io-core-adapter-fastify@0.0.6
+
 ## 0.1.0
 
 ### Minor Changes