+++ packages/io-core-adapter-fims/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pagopa/io-core-adapter-fims
 
+## 0.0.6
+
+### Patch Changes
+
+- f146ba6: fixed endpoints with /api
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [1e199aa]
+  - @pagopa/io-core-adapter-fastify@0.0.6
+
 ## 0.0.4
 
 ### Patch Changes