+++ packages/io-core-adapter-tracing/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pagopa/io-core-adapter-tracing
 
+## 0.0.6
+
+### Patch Changes
+
+- 89fa17c: Added error tracing
+
+## 0.0.5
+
+### Patch Changes
+
+- 13a5a52: fixed monitor init
+
 ## 0.0.4
 
 ### Patch Changes